A Cabin Fever Challenge target generator
The generator supports arbitary distances, between
a max and min distance, and will scale the target appropriately
generating a pdf file for either Letter or A4 paper.

If this becomes offical, it should be hooked up to a web form
for the user to just select divsion, distance, and paper to
get what they need.